thematic_analysis/requirements.txt
python-dotenv>=1.0.0
pdfplumber>=0.10.3
pandas>=2.0.0
requests>=2.31.0
openpyxl>=3.1.2 